Improvements to dialog parsing + debugged renderChatMessage hook
There appears to have been a subtle bug in suppressVisited(). When a renderChatMessage hook fires, regardless of whether it is the first or the fifteenth call, the document is not updated with previous 'rendered' chat messages. As such, suppressVisited must use the hook's html field as a parent DOM node during that stage of Foundry bootup